refactor(webapp): own plan visibility in one module
Both plan classifications now live in planVisibility.ts, renamed from
legacyPlans.ts: which plans are legacy, for the usage banner, and
which get a summary strip. They answer different questions — a
free-uncapped account is on a current plan but still shows no strip —
so they stay separate maps, but a reader looking for "which plans are
what" now has one place to look. Both are exhaustive over the plan
name union, so a new plan fails to compile until it is classified for
both.

Also hide the strip when the plans list fails to load. Titles come
from that endpoint, so without it the strip could only print raw Orb
codes like "growth-v2"; the previous guard checked for pending only,
which is false once a request has errored.
